+++ packages/zsa-react/src/index.ts
@@ -59,8 +59,9 @@ export const useServerAction = <
 
   // keep track of pending states
   const [isTransitioning, startTransition] = useTransition()
-  const [status, setStatus] =
-    useState<TServerActionResult<TServerAction>["status"]>("idle")
+  const [isExecuting, setExecuting] = useState(false)
+
+  const status = useRef<TServerActionResult<TServerAction>["status"]>("idle")
 
   // set the result state and ref
   const setResult = useCallback((result: TInnerResult<TServerAction>) => {
@@ -108,7 +109,8 @@ export const useServerAction = <
 
       if (opts?.onStart) opts.onStart()
 
-      setStatus("pending")
+      status.current = "pending"
+      setExecuting(true)
 
       let data, err
 
@@ -135,8 +137,6 @@ export const useServerAction = <
           )
         }
 
-        setStatus("error")
-
         // don't retry => update the result
         if (oldResult.status === "filled") {
           setResult(oldResult.result)
@@ -150,11 +150,12 @@ export const useServerAction = <
           result: undefined,
         })
 
+        // trigger error useEffect
+        status.current = "error"
+        setExecuting(false)
+
         return [data, err] as any
       }
-
-      // success state
-      setStatus("success")
 
       const res = {
         isError: false,
@@ -169,6 +170,10 @@ export const useServerAction = <
         status: "empty",
         result: undefined,
       })
+
+      // success state
+      status.current = "success"
+      setExecuting(false)
 
       return [data, err] as any
     },
@@ -186,6 +191,22 @@ export const useServerAction = <
         startTransition(() => {
           internalExecute(opts[0])
         })
+      })
+    },
+    [internalExecute]
+  )
+
+  const executeFormAction = useCallback(
+    async (
+      ...opts: Parameters<TServerAction>[0] extends undefined
+        ? []
+        : [Parameters<TServerAction>[0]]
+    ): Promise<null> => {
+      return await new Promise((resolve) => {
+        startTransition(() => {
+          internalExecute(opts[0])
+        })
+        resolve(null)
       })
     },
     [internalExecute]
@@ -218,25 +239,24 @@ export const useServerAction = <
   const reset = useCallback(() => {
     setResult(getEmptyResult())
     setOldResult(getEmptyOldResult())
-    setStatus("idle")
+    setExecuting(false)
+    status.current = "idle"
     lastRetryId.current = 0
     retryCount.current = 0
   }, [])
 
-  // check if the status is pending
-  // NOTE: during retries, the status is "pending" but the transition is not
-  //       complete, so we need to check the status to see if it is pending
-  const isPending = isTransitioning || status === "pending"
+  const isRunningCallbacks = useRef(false)
 
-  useEffect(() => {
-    // we need this effect because we won't know when the next.js server action is
-    // actually done until the transition finishes and sets isTransitioning back to false.
-    // when the transition finishes, we call resolve the executeRef.current promise and also
-    // invoke the onSuccess and onError callbacks.
-    if (isPending) return
+  const handleCallbacks = useCallback(() => {
+    if (!executeRef.current || isRunningCallbacks.current) {
+      return
+    }
+
+    // make sure we don't call this function multiple times
+    isRunningCallbacks.current = true
 
     // handle the success state
-    if (status === "success") {
+    if (status.current === "success") {
       executeRef.current?.([resultRef.current.data, null])
 
       // call success callback
@@ -249,7 +269,7 @@ export const useServerAction = <
     }
 
     // handle the error state
-    if (status === "error") {
+    if (status.current === "error") {
       executeRef.current?.([null, resultRef.current.error])
 
       // call error callback
@@ -260,7 +280,40 @@ export const useServerAction = <
       // call finish callback
       opts?.onFinish?.([null, resultRef.current.error] as any)
     }
-  }, [status, isPending])
+
+    // reset the states
+    executeRef.current = undefined
+    status.current = "idle"
+    isRunningCallbacks.current = false
+  }, [])
+
+  // check if the status is pending
+  // NOTE: during retries, the status is "pending" but the transition is not
+  //       complete, so we need to check the status to see if it is pending
+  const isPending = isTransitioning || isExecuting
+
+  useEffect(() => {
+    // we need this effect because we won't know when the next.js server action is
+    // actually done until the transition finishes and sets isTransitioning back to false.
+    // when the transition finishes, we call resolve the executeRef.current promise and also
+    // invoke the onSuccess and onError callbacks.
+    if (isPending) return
+
+    // handle the callbacks
+    handleCallbacks()
+  }, [status.current, isPending])
+
+  // on a revalidatePath or redirect the isPending useEffect won't run
+  // to combat this, we need to have a cleanup method that will fire
+  // when the hook is unmounted. we will run the callbacks here
+  // since they won't be called from the useEffect
+  useEffect(() => {
+    return () => {
+      if (executeRef.current !== undefined) {
+        handleCallbacks()
+      }
+    }
+  }, [])
 
   const final = calculateResultFromState<TServerAction>({
     isPending,
@@ -273,5 +326,6 @@ export const useServerAction = <
     reset,
     execute,
     setOptimistic,
+    executeFormAction,
   }
 }